requirements.txt
asgiref==3.11.1
certifi==2026.2.25
Django==6.0.4
django-jazzmin==3.0.4
gunicorn==25.3.0
packaging==26.1
pillow==12.2.0
python-dotenv==1.2.2
sqlparse==0.5.5
whitenoise==6.12.0
